BitsX.jl

BitsX

BitsX — Module

module BitsX

BitsX is a collection of types and functions for working with representations bits.

A main idea is to have a convenient, performant, and uniform interface to working with representations of bits in various data types. For example a vector of bits may be represented as a value of any of Vector{Bool}, UInt64, String, etc. The abstractions supporting this idea are in BitsBase (for the most part).

Vocabulary

bit string: an AbstractString whose characters are all '1' and '0'. In some contexts, formatting characters may be present as well.

Warn

The API of BitsX is large, perhapse too large. Much of it is rather exploratory and incomplete. In particular, interfaces for types that require a lot of implemenation are incomplete.

BitsX.jl has recently been reorganized and modularized. It might be a good idea to split it into smaller packages.

Related packages

BitBasis.jl

BitFlags.jl

BitInformation.jl

BitIntegers.jl

BitOperations.jl

BitPermutations.jl

Bits.jl

FieldFlags.jl

BitTwiddlingConvenienceFunctions

BitConverter.jl

BitVectorExtensions.jl

BitMasks.jl

FlagSets.jl

BitsFields.jl

BitSetTuples.jl

BitFloats.jl

Contents

BitsX.jl

BitsX

Contents

Index

Contents

BitsBase

binzero

binone

isbinzero

isbinzero

isbin

is_zero_char

is_one_char

is_binary_char

to_binary_char

to_binary_char_code

from_binary_char

bitconvert

min_bits

min_dits

bitaxes

biteachindex

bitlastindex

bitsizeof

bitsize

bitlength

count_bits

bit

tstbit

bit0

tstbit0

ZeroBased

OneBased

Contents

Creating and parsing bit strings

Functions

parse_bin

bstring

bstringview

normalize_bitstring

randbitstring

randbitstring!

is_bitstring

check_bitstring

Contents

Masking

mask

masked

rightmask

leftmask

rangemask

Miscellany

undigits

bitcollect

bit_count_ones

bit_count_zeros

BitStringArray

Everything else

Index

BitsX

BitsX.BitArrays

BitsX.BitsBase

BitsX.BitArrayViews.BitArrayView

BitsX.BitArraysX._BitArraysX.Chunks

BitsX.Bits.BitStringArray

BitsX.BitsBase.OneBased

BitsX.BitsBase.ZeroBased

BitsX.BStringViews.bstringview

BitsX.BStrings.bstring

BitsX.BitArrayViews.bitmatview

BitsX.BitArrayViews.bitvecview

BitsX.BitArrayViews.is_little_endian

BitsX.BitArrays.bitarray

BitsX.BitArrays.bitarray!

BitsX.BitArrays.bitvector

BitsX.BitIntegersX.get_one_bit_masks

BitsX.BitIntegersX.min_uint_bit_width

BitsX.BitIntegersX.min_uint_byte_width

BitsX.BitIntegersX.min_uint_type

BitsX.BitIntegersX.uint_type

BitsX.BitIntegersX.uint_type_bytes

BitsX.Bits.bit_count_ones

BitsX.Bits.bit_count_zeros

BitsX.Bits.bitcollect

BitsX.Bits.leftmask

BitsX.Bits.mask

BitsX.Bits.masked

BitsX.Bits.normalize_bitstring

BitsX.Bits.randbitstring

BitsX.Bits.randbitstring!

BitsX.Bits.rangemask

BitsX.Bits.rightmask

BitsX.Bits.undigits

BitsX.BitsBase.binone

BitsX.BitsBase.binzero

BitsX.BitsBase.bit

BitsX.BitsBase.bit0

BitsX.BitsBase.bitaxes

BitsX.BitsBase.bitconvert

BitsX.BitsBase.biteachindex

BitsX.BitsBase.bitlastindex

BitsX.BitsBase.bitlength

BitsX.BitsBase.bitsize

BitsX.BitsBase.bitsizeof

BitsX.BitsBase.check_bitstring

BitsX.BitsBase.count_bits

BitsX.BitsBase.from_binary_char

BitsX.BitsBase.is_binary_char

BitsX.BitsBase.is_bitstring

BitsX.BitsBase.is_one_char

BitsX.BitsBase.is_zero_char

BitsX.BitsBase.isbin

BitsX.BitsBase.isbinone

BitsX.BitsBase.isbinzero

BitsX.BitsBase.min_bits

BitsX.BitsBase.min_dits

BitsX.BitsBase.to_binary_char

BitsX.BitsBase.to_binary_char_code

BitsX.BitsBase.tstbit

BitsX.BitsBase.tstbit0

BitsX.ParseBin.parse_bin